refactor gtc to gtc2, was shadowed by eve-gtc
