+++ src/gtc2/common.py
