+++ src/gtc2/gtir.py
