+++ tests/test_unittest/test_cli.py
@@ -78,7 +78,7 @@ BACKEND_ROW_PATTERN_BY_NAME = {
     "dawn:naive": r"^\s*dawn:naive\s*c\+\+\s*python\s*No",
     "dawn:cxxopt": r"^\s*dawn:cxxopt\s*c\+\+\s*python\s*No",
     "dawn:cuda": r"^\s*dawn:cuda\s*cuda\s*python\s*No",
-    "gtc:py": r"^\s*gtc:py\s*python\s*python\s*Yes",
+    "gtc2:py": r"^\s*gtc2:py\s*python\s*python\s*Yes",
     "nocli": r"^\s*nocli\s*\?\s*\?\s*No",
 }
